+++ src/components/PrimaryButton/index.ts
@@ -1,0 +1,2 @@
+export { PrimaryButton } from "@/components/PrimaryButton/PrimaryButton";
+export type { PrimaryButtonProps } from "@/components/PrimaryButton/interfaces";