+++ src/components/PrimaryButton/interfaces.ts
@@ -1,0 +1,13 @@
+export type ButtonTooltip = {
+  whenDisabled?: string;
+  whenEnabled?: string;
+};
+
+export type PrimaryButtonProps = {
+  text: string;
+  disabled?: boolean;
+  tooltip?: ButtonTooltip;
+  onClick?: () => void;
+  type?: "button" | "submit" | "reset";
+  className?: string;
+};