+++ src/components/SecondaryButton/index.ts
@@ -1,0 +1,2 @@
+export { SecondaryButton } from "@/components/SecondaryButton/SecondaryButton";
+export type { SecondaryButtonProps } from "@/components/SecondaryButton/interfaces";