+++ src/components/Toast/index.ts
@@ -1,0 +1,1 @@
+export { catToast, Toaster } from "@/components/Toast/Toast";