+++ src/modules/options-dashboard/views/CatProfileDashboardView/index.ts
@@ -1,0 +1,2 @@
+export { CatProfileDashboardView } from "@/modules/options-dashboard/views/CatProfileDashboardView/CatProfileDashboardView";
+export type { CatProfileDashboardViewProps } from "@/modules/options-dashboard/views/CatProfileDashboardView/interfaces";